src/frontend/requirements.txt
pandas
requests
streamlit
python-dotenv
